+++ day6/day6/src/main/java/org/sccmckenzie/aoc2023/day6/RaceRecord.java
@@ -6,8 +6,8 @@ import lombok.Getter;
 @AllArgsConstructor
 public class RaceRecord {
     @Getter
-    private int time;
+    private long time;
 
     @Getter
-    private int distance;
+    private long distance;
 }
